Convert AFNI File to NIFTI Image

Converts an AFNI file to an NIfTI file or nifti using 3dTshift

afni_3dTshift(file, tr, tpattern = c("alt+z", "alt+z2", "alt-z",
  "alt-z2", "seq+z", "seq-z"), tzero = NULL, slice = NULL,
  ignore_vols = 0, opts = "", retimg = TRUE, ...)

tshift(...)

Arguments

file
AFNI BRIK or HEAD file

tr
Repetition time. You may attach the suffix 's' for seconds, or 'ms' for milliseconds.

tpattern
Pattern of slice time pattern. alt+z = altplus = alternating in the plus direction alt+z2 = alternating, starting at slice #1 instead of #0 alt-z = altminus = alternating in the minus direction alt-z2 = alternating, starting at slice #nz-2 instead of #nz-1 seq+z = seqplus = sequential in the plus direction seq-z = seqminus = sequential in the minus direction

tzero
time offset to align each slice to (The default alignment time is the average of the 'tpattern' values)

slice
slice number indicating the time offset to align each slice to. If this is specified tzero may not be, and vice versa.

ignore_vols
Number of volumes to ignore/delete

opts
Additional options to pass to 3dTshift

retimg
Should an image be returned?

...
additional arguments to readnii

Value

If retimg then object of class nifti. Otherwise, the outfile argument is returned (or a temporary filename)