AFNI 3dresample function

Wrapper for AFNI 3dresample function

afni_3dresample(file, bound_type = NULL, dxyz = NULL, orient = NULL,
  resample_mode = c("NN", "Li", "Cu", "Bk"), master = NULL,
  outfile = NULL, retimg = TRUE, opts = "", ...)

resample(...)

Arguments

file

nifti object or NIfTI filename.

bound_type

specify which boundary is preserved, options are "FOV" (default if NULL) or "SLAB"

dxyz

resample to new dimensions for x, y, and z-directions

orient

new axis order. The orientation code is a 3 character string, where the characters come from the respective sets A,P, I,S, L,R. For example LPI makes the x-axis is Left-to-Right, the y-axis is Posterior-to-Anterior, and the z-axis is Inferior-to-Superior.

resample_mode

Resampling method. The options are for 'Nearest Neighbor', 'Linear', 'Cubic' and 'Blocky' interpolation, respectively.

master

a filename of an image, which the input file is aligned to (not doing registration, but takes orientation and pixel dimension information)

outfile

Output filename (should not have an extension)

retimg

Should an image be returned (TRUE) or a filename?

opts

Additional options passed to 3dresample

...

additional arguments to afni_3dAFNItoNIFTI

Value

Output filename of the brik