afni Command Wrapper

This function calls AFNI command passed to func

afni_cmd(func, file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, opts = "", verbose = TRUE, samefile = FALSE,
  opts_after_outfile = FALSE, frontopts = "", add_ext = TRUE,
  bin_app = "bin", quote_file = TRUE, quote_outfile = TRUE,
  run = TRUE, ...)

Arguments

func

(character) AFNI function

file

(character) image to be manipulated

outfile

(character) resultant image name (optional)

retimg

(logical) return image of class nifti

reorient

(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern

(logical) to be passed to system

opts

(character) operations to be passed to func

verbose

(logical) print out command before running

samefile

(logical) is the output the same file?

opts_after_outfile

(logical) should opts come after the outfile in the AFNI command?

frontopts

(character) options/character to put in before filename

add_ext

(logical) should the extension be added to the outfile

bin_app

(character) appendix to add to get_afni

quote_file

should the filename be quoted in double quotes?

quote_outfile

should the output filename be quoted in double quotes?

run

(logical) Should the command be run? If this is false, nothing is done

...

additional arguments passed to system.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.